@@ -6,7 +6,8 @@
   "main": "dist/index.js",
   "typings": "src/index.ts",
   "scripts": {
-    "_setup": "npm i -g commitizen semantic-release-cli",
+    "dev": "npm run server",
+    "server": "webpack-dev-server --hot --colors",
     "t": "npm run mocha -- src/*.spec.ts",
     "test": "npm run eslint && npm run tscheck && npm run t",
     "test:r": "npm run mocha -- src/*.spec.ts",
@@ -20,33 +21,49 @@
     "eslint": "eslint src --ext ts",
     "eslint:fix": "eslint src --ext \"ts\" --fix",
     "eslint:w": "chokidar \"src/**/*.ts\" --initial -c \"npm run eslint\"",
-    "lint:w": "concurrently -- \"npm run tsclint:w\" \"npm run eslint:w\"",
-    "build": "tsc",
-    "prebuild": "npm run clean",
+    "lint:w": "run-p tsclint:w eslint:w",
+    "build:libs": "npm run build -- --env.libs --progress",
+    "build": "webpack",
     "clean": "rimraf dist",
     "commit": "git-cz"
   },
   "dependencies": {
-    "@hyperapp/html": "^1.1.1",
-    "hyperapp": "^1.2.6",
-    "tslib": "^1.9.0"
+    "@hyperapp/html": "1.1.1",
+    "hyperapp": "1.2.6",
+    "tslib": "1.9.2"
   },
   "devDependencies": {
-    "@types/mocha": "^5.2.0",
-    "@types/node": "^10.0.6",
-    "chokidar-cli": "^1.2.0",
-    "concurrently__replace__by_npm_run_all": "^3.5.1",
-    "eslint": "^4.19.1",
-    "eslint-plugin-import": "^2.11.0",
-    "eslint-plugin-tslint": "^2.1.0",
-    "eslint-plugin-typescript": "^0.12.0",
-    "eslint-plugin-unicorn": "^4.0.3",
-    "mocha": "^5.1.1",
-    "rimraf": "^2.6.2",
-    "ts-node": "^6.0.3",
-    "tslint": "^5.10.0",
-    "typescript": "2.8.3",
-    "typescript-eslint-parser": "^15.0.0"
+    "1-liners": "0.4.0",
+    "@types/execa": "0.9.0",
+    "@types/mocha": "5.2.1",
+    "@types/node": "10.3.1",
+    "@types/webpack": "4.4.0",
+    "@types/webpack-env": "1.13.6",
+    "add-asset-html-webpack-plugin": "2.1.3",
+    "autoprefixer": "8.6.0",
+    "chokidar-cli": "1.2.0",
+    "ejs-loader": "0.3.1",
+    "eslint": "4.19.1",
+    "eslint-plugin-import": "2.12.0",
+    "eslint-plugin-tslint": "2.1.0",
+    "eslint-plugin-typescript": "0.12.0",
+    "eslint-plugin-unicorn": "4.0.3",
+    "execa": "0.10.0",
+    "extract-text-webpack-plugin": "3.0.2",
+    "html-webpack-plugin": "3.2.0",
+    "mocha": "5.2.0",
+    "npm-run-all": "4.1.3",
+    "rimraf": "2.6.2",
+    "script-ext-html-webpack-plugin": "2.0.1",
+    "ts-loader": "4.3.1",
+    "ts-node": "5.0.1",
+    "tslint": "5.10.0",
+    "typescript": "2.9.1",
+    "typescript-eslint-parser": "16.0.0",
+    "webpack": "4.11.0",
+    "webpack-cli": "3.0.2",
+    "webpack-dev-server": "3.1.4",
+    "webpack-loader-helper": "1.0.0"
   },
   "engines": {
     "node": ">=6",
