@@ -16,6 +16,11 @@ const tslintRules = Object.assign({}, tslintRulesRecommended, {
     "interface-over-type-literal": false,
     "align": false,
     "trailing-comma": false,
+    "no-string-literal": false,
+    "no-namespace": false,
+    "interface-name": false,
+    "no-angle-bracket-type-assertion": false,
+    "whitespace": false,
 });
 
 module.exports = {
@@ -31,7 +36,10 @@ module.exports = {
     "parser": "typescript-eslint-parser",
     "parserOptions": {
         "ecmaVersion": 2017,
-        "sourceType": "module"
+        "sourceType": "module",
+        "ecmaFeatures": {
+            "jsx": true
+        },
     },
     "plugins": [
         "unicorn",
@@ -44,10 +52,11 @@ module.exports = {
         "no-unused-vars": 0,
         "indent": 0,
         "unicorn/import-index": 0,
-        "tslint/config": [1, {
-            rules: tslintRules,
-            rulesDirectory: ["node_modules/tslint/lib/rules"],
-        }],
+        "no-console": 1,
+        // "tslint/config": [1, {
+        //     rules: tslintRules,
+        //     rulesDirectory: ["node_modules/tslint/lib/rules"],
+        // }],
         "import/newline-after-import": 0,
         "import/no-duplicates": 1,
         "import/max-dependencies": [1, { "max": 10 }],
